+++ client.go
@@ -47,9 +47,9 @@ type Client struct {
 	RefreshToken string
 	// UserAgent is the HTTP User-Agent string
 	UserAgent string
-	// Usr is the FMC username.
+	// Usr is the FMC username. Not used for cdFMC.
 	Usr string
-	// Pwd is the FMC password.
+	// Pwd is the FMC password or cdFMC API token
 	Pwd string
 	// Insecure determines if insecure https connections are allowed.
 	Insecure bool
@@ -67,12 +67,14 @@ type Client struct {
 	LastRefresh time.Time
 	// RefreshCount is the number to authentication token refreshes with the same refresh token
 	RefreshCount int
-	// DomainUUID is the UUID of the global domain returned when generating a token
+	// DomainUUID is the UUID of the user login domain.
 	DomainUUID string
-	// Map of domain names to domain UUIDs
+	// Map of domain names to domain UUIDs. Not applicable for cdFMC
 	Domains map[string]string
 	// FMC Version
 	FMCVersion string
+	// Is this cdFMC connection
+	IsCDFMC bool
 
 	RateLimiterBucket *ratelimit.Bucket
 
@@ -119,6 +121,32 @@ func NewClient(url, usr, pwd string, mods ...func(*Client)) (Client, error) {
 	err := client.GetFMCVersion()
 	if err != nil {
 		return client, err
+	}
+
+	return client, nil
+}
+
+// Create a new cdFMC HTTP client.
+func NewClientCDFMC(url, apiToken string, mods ...func(*Client)) (Client, error) {
+	// Set client mode to cdFMC
+	mods = append(mods, cdFMC(true))
+
+	// Create client as usual. Username is not used.
+	client, err := NewClient(url, "", apiToken, mods...)
+	if err != nil {
+		return client, err
+	}
+
+	// Get the Global Domain UUID. cdFMC does not support multi-domain.
+	// Global UUID is fixed (e276abec-e0f2-11e3-8169-6d9ed49b625f), though we get it from the cdFMC just in case.
+	res, err := client.Get("/api/fmc_platform/v1/info/domain")
+	if err != nil {
+		return client, err
+	}
+	if uuid := res.Get("items.0.uuid"); !uuid.Exists() {
+		return client, fmt.Errorf("failed to retrieve domain UUID from: %s", res.String())
+	} else {
+		client.DomainUUID = uuid.String()
 	}
 
 	return client, nil
@@ -180,6 +208,13 @@ func BackoffDelayFactor(x float64) func(*Client) {
 	}
 }
 
+// cdFMC sets connector mode to cdFMC (true) or FMC (false).
+func cdFMC(x bool) func(*Client) {
+	return func(client *Client) {
+		client.IsCDFMC = x
+	}
+}
+
 // NewReq creates a new Req request for this client.
 // Use a "{DOMAIN_UUID}" placeholder in the URI to be replaced with the domain UUID.
 func (client Client) NewReq(method, uri string, body io.Reader, mods ...func(*Req)) (Req, error) {
@@ -218,7 +253,11 @@ func (client Client) NewReq(method, uri string, body io.Reader, mods ...func(*Re
 //	res, _ := client.Do(req)
 func (client *Client) Do(req Req) (Res, error) {
 	// add token
-	req.HttpReq.Header.Add("X-auth-access-token", client.AuthToken)
+	if client.IsCDFMC {
+		req.HttpReq.Header.Add("Authorization", "Bearer "+client.Pwd)
+	} else {
+		req.HttpReq.Header.Add("X-auth-access-token", client.AuthToken)
+	}
 	req.HttpReq.Header.Add("Content-Type", "application/json")
 	req.HttpReq.Header.Add("Accept", "application/json")
 	req.HttpReq.Header.Add("User-Agent", client.UserAgent)
@@ -271,7 +310,7 @@ func (client *Client) Do(req Req) (Res, error) {
 			} else if httpRes.StatusCode == 429 || (httpRes.StatusCode >= 500 && httpRes.StatusCode <= 599) {
 				log.Printf("[ERROR] HTTP Request failed: StatusCode %v, Retries: %v", httpRes.StatusCode, attempts)
 				continue
-			} else if httpRes.StatusCode == 401 {
+			} else if httpRes.StatusCode == 401 && !client.IsCDFMC {
 				// There are bugs in FMC, where the sessions are invalidated out of the blue
 				// In case such a situation is detected, new authentication is forced
 				log.Printf("[DEBUG] Invalid session detected. Forcing reauthentication")
@@ -529,6 +568,11 @@ func (client *Client) Refresh() error {
 
 // Login if no token available.
 func (client *Client) Authenticate() error {
+	// cdFMC uses fixed token, no need to do separate authentication
+	if client.IsCDFMC {
+		return nil
+	}
+
 	var err error
 	client.authenticationMutex.Lock()
 	// Check if we can attempt to refresh the token (there is old token, it's between 25 and 29 minutes since last refresh, and less than 3 refreshes done)